+++ lib/Sema/TypeCheckGeneric.cpp
@@ -16,6 +16,7 @@
 #include "TypeChecker.h"
 #include "GenericTypeResolver.h"
 #include "swift/AST/ArchetypeBuilder.h"
+#include "swift/Basic/Defer.h"
 
 using namespace swift;
 
@@ -781,19 +782,179 @@ GenericSignature *TypeChecker::validateGenericSignature(
   return sig;
 }
 
+static void revertDependentTypeLoc(TypeLoc &tl) {
+  // If there's no type representation, there's nothing to revert.
+  if (!tl.getTypeRepr())
+    return;
+
+  // Don't revert an error type; we've already complained.
+  if (tl.wasValidated() && tl.isError())
+    return;
+
+  // Make sure we validate the type again.
+  tl.setType(Type(), /*validated=*/false);
+}
+
+/// Finalize the given generic parameter list, assigning archetypes to
+/// the generic parameters.
+void TypeChecker::finalizeGenericParamList(ArchetypeBuilder &builder,
+                                           GenericParamList *genericParams,
+                                           DeclContext *dc) {
+  Accessibility access;
+  if (auto *fd = dyn_cast<FuncDecl>(dc))
+    access = fd->getFormalAccess();
+  else if (auto *nominal = dyn_cast<NominalTypeDecl>(dc))
+    access = nominal->getFormalAccess();
+  else
+    access = Accessibility::Internal;
+
+  // Wire up the archetypes.
+  for (auto GP : *genericParams) {
+    GP->setArchetype(builder.getArchetype(GP));
+    checkInheritanceClause(GP);
+    if (!GP->hasAccessibility())
+      GP->setAccessibility(access);
+  }
+  genericParams->setAllArchetypes(
+    Context.AllocateCopy(builder.getAllArchetypes()));
+
+#ifndef NDEBUG
+  // Record archetype contexts.
+  for (auto archetype : genericParams->getAllArchetypes()) {
+    if (Context.ArchetypeContexts.count(archetype) == 0)
+      Context.ArchetypeContexts[archetype] = dc;
+  }
+#endif
+
+  // Replace the generic parameters with their archetypes throughout the
+  // types in the requirements.
+  // FIXME: This should not be necessary at this level; it is a transitional
+  // step.
+  for (auto &Req : genericParams->getRequirements()) {
+    if (Req.isInvalid())
+      continue;
+
+    switch (Req.getKind()) {
+    case RequirementReprKind::TypeConstraint: {
+      revertDependentTypeLoc(Req.getSubjectLoc());
+      if (validateType(Req.getSubjectLoc(), dc)) {
+        Req.setInvalid();
+        continue;
+      }
+
+      revertDependentTypeLoc(Req.getConstraintLoc());
+      if (validateType(Req.getConstraintLoc(), dc)) {
+        Req.setInvalid();
+        continue;
+      }
+      break;
+    }
+
+    case RequirementReprKind::SameType:
+      revertDependentTypeLoc(Req.getFirstTypeLoc());
+      if (validateType(Req.getFirstTypeLoc(), dc)) {
+        Req.setInvalid();
+        continue;
+      }
+
+      revertDependentTypeLoc(Req.getSecondTypeLoc());
+      if (validateType(Req.getSecondTypeLoc(), dc)) {
+        Req.setInvalid();
+        continue;
+      }
+      break;
+    }
+  }
+}
+
+/// Revert the dependent types within the given generic parameter list.
+void TypeChecker::revertGenericParamList(GenericParamList *genericParams) {
+  // Revert the inherited clause of the generic parameter list.
+  for (auto param : *genericParams) {
+    param->setCheckedInheritanceClause(false);
+    for (auto &inherited : param->getInherited())
+      revertDependentTypeLoc(inherited);
+  }
+
+  // Revert the requirements of the generic parameter list.
+  for (auto &req : genericParams->getRequirements()) {
+    if (req.isInvalid())
+      continue;
+
+    switch (req.getKind()) {
+    case RequirementReprKind::TypeConstraint: {
+      revertDependentTypeLoc(req.getSubjectLoc());
+      revertDependentTypeLoc(req.getConstraintLoc());
+      break;
+    }
+
+    case RequirementReprKind::SameType:
+      revertDependentTypeLoc(req.getFirstTypeLoc());
+      revertDependentTypeLoc(req.getSecondTypeLoc());
+      break;
+    }
+  }
+}
+
 bool TypeChecker::validateGenericTypeSignature(GenericTypeDecl *typeDecl) {
   bool invalid = false;
-  if (!typeDecl->isValidatingGenericSignature()) {
-    typeDecl->setIsValidatingGenericSignature();
-    auto sig = validateGenericSignature(typeDecl->getGenericParams(),
-                                        typeDecl->getDeclContext(),
-                                        nullptr, nullptr, invalid);
-    assert(sig->getInnermostGenericParams().size()
-             == typeDecl->getGenericParams()->size());
-    typeDecl->setGenericSignature(sig);
-    typeDecl->setIsValidatingGenericSignature(false);
+
+  if (typeDecl->isValidatingGenericSignature())
+    return invalid;
+
+  typeDecl->setIsValidatingGenericSignature();
+
+  defer { typeDecl->setIsValidatingGenericSignature(false); };
+
+  auto *gp = typeDecl->getGenericParams();
+  auto *dc = typeDecl->getDeclContext();
+
+  auto sig = validateGenericSignature(gp, dc, nullptr, nullptr, invalid);
+  assert(sig->getInnermostGenericParams().size()
+           == typeDecl->getGenericParams()->size());
+  typeDecl->setGenericSignature(sig);
+
+  if (invalid) {
+    markInvalidGenericSignature(typeDecl);
+    return invalid;
   }
+
+  revertGenericParamList(gp);
+
+  ArchetypeBuilder builder =
+    createArchetypeBuilder(typeDecl->getModuleContext());
+  auto *parentSig = dc->getGenericSignatureOfContext();
+  checkGenericParamList(&builder, gp, parentSig);
+  finalizeGenericParamList(builder, gp, typeDecl);
+
   return invalid;
+}
+
+void TypeChecker::revertGenericFuncSignature(AbstractFunctionDecl *func) {
+  // Revert the result type.
+  if (auto fn = dyn_cast<FuncDecl>(func))
+    if (!fn->getBodyResultTypeLoc().isNull())
+      revertDependentTypeLoc(fn->getBodyResultTypeLoc());
+
+  // Revert the body parameter types.
+  for (auto paramList : func->getParameterLists()) {
+    for (auto &param : *paramList) {
+      // Clear out the type of the decl.
+      if (param->hasType() && !param->isInvalid())
+        param->overwriteType(Type());
+      revertDependentTypeLoc(param->getTypeLoc());
+    }
+  }
+
+  // Revert the generic parameter list.
+  if (func->getGenericParams())
+    revertGenericParamList(func->getGenericParams());
+
+  // Clear out the types.
+  if (auto fn = dyn_cast<FuncDecl>(func))
+    fn->revertType();
+  else
+    func->overwriteType(Type());
 }
 
 /// Create a text string that describes the bindings of generic parameters that
